refactor: added use-page-state hook for onboarding
